Replace Double with BigDecimal in financial models

+++ src/main/java/com/AIT/Optimanage/Analytics/DTOs/PrevisaoDTO.java
@@ -4,10 +4,12 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class PrevisaoDTO {
-    private double valorPrevisto;
+    private BigDecimal valorPrevisto;
 }
